requirements.txt
streamlit
numpy
plotly
Pillow
pytest